+++ src/lint/src/rules/gml/rules/normalize-doc-param-undefined-defaults-rule.ts
@@ -1,7 +1,7 @@
 import type { Rule } from "eslint";
 
 import type { GmlRuleDefinition } from "../index.js";
-import { createMeta, resolveLocFromIndex } from "../rule-base-helpers.js";
+import { createMeta, reportLineTextFixes } from "../rule-base-helpers.js";
 
 const undefinedOptionalDefaultPattern =
     /^(\s*\/\/\/\s*@param(?:\s+\{[^}\r\n]+\})?\s+)\[([A-Za-z0-9_]+)\s*=\s*undefined\](.*)$/u;
@@ -13,34 +13,6 @@ function normalizeUndefinedOptionalDefaultLine(line: string): string {
     }
 
     return `${normalized[1]}[${normalized[2]}]${normalized[3]}`;
-}
-
-type SourceLine = Readonly<{
-    startOffset: number;
-    text: string;
-}>;
-
-function collectSourceLines(sourceText: string): ReadonlyArray<SourceLine> {
-    const lines: Array<SourceLine> = [];
-    const linePattern = /[^\r\n]*(?:\r\n|\r|\n|$)/gu;
-    let match: RegExpExecArray | null;
-    while ((match = linePattern.exec(sourceText)) !== null) {
-        const rawLine = match[0];
-        if (rawLine.length === 0 && match.index === sourceText.length) {
-            break;
-        }
-
-        lines.push({
-            startOffset: match.index,
-            text: rawLine.replace(/(?:\r\n|\r|\n)$/u, "")
-        });
-
-        if (linePattern.lastIndex === sourceText.length) {
-            break;
-        }
-    }
-
-    return lines;
 }
 
 /**
@@ -55,23 +27,6 @@ export function sanitizeDocCommentUndefinedOptionalParamDefaults(text: string): 
         .split(/\r?\n/u)
         .map((line) => normalizeUndefinedOptionalDefaultLine(line))
         .join(lineEnding);
-}
-
-function reportUndefinedOptionalDefaultFixes(context: Rule.RuleContext, definition: GmlRuleDefinition): void {
-    const sourceText = context.sourceCode.text;
-    for (const line of collectSourceLines(sourceText)) {
-        const normalizedLine = normalizeUndefinedOptionalDefaultLine(line.text);
-        if (normalizedLine === line.text) {
-            continue;
-        }
-
-        context.report({
-            loc: resolveLocFromIndex(context, sourceText, line.startOffset),
-            messageId: definition.messageId,
-            fix: (fixer) =>
-                fixer.replaceTextRange([line.startOffset, line.startOffset + line.text.length], normalizedLine)
-        });
-    }
 }
 
 /**
@@ -89,7 +44,7 @@ export function createNormalizeDocParamUndefinedDefaultsRule(definition: GmlRule
         create(context: Rule.RuleContext): Rule.RuleListener {
             return {
                 Program() {
-                    reportUndefinedOptionalDefaultFixes(context, definition);
+                    reportLineTextFixes(context, definition, normalizeUndefinedOptionalDefaultLine);
                 }
             };
         }
